facebook-mqtt: fixed preemptive connection timeouts
As it stands, a connection is declared as being timed out if it has not
heard back from the server with a ping response in one keep-alive time
interval. However, the MQTT specification states that a connection is
only declared timed out after one and a half time intervals. While the
effects of this oversight may not be immediately present, over a period
uptime, the connection may fall victim to preemptive timeouts.