facebook: ensure the account is online with root commands
Currently, it is just assumed that the account is online, however, this
is not always the case. If one of the root commands attempts to act on
an offline account, it will typically lead to a segmentation fault. To
guard against this, an additional check must be added to ensure the
account is actually online.