facebook: default to chat_type of room for persistence
As of now, group chat channels do not persist after they are left. This
is due in part to the channel not being of chat_type room. Furthermore,
the account and room identifiers are not set on the channel.

This issue becomes apparent when the plugin disconnects from the remote
services for whatever reason, and the channel is not setup again.

This issue was introduced in 06956d1.