fb_mqtt_write: fix unreachable input_add (priv->wev was always 0)
This was the only line that assigned anything to priv->wev, and it was
behind an (incorrect) condition that checked if it's nonzero.

That would have replaced priv->wev if the condition was ever true, but
since it wasn't, the only result is potentially delayed writes
(for example, filling the write buffer and only writing the rest a
minute later when pinging the server)

The new condition also checks the return value of fb_mqtt_cb_write(),
which is true if it should continue writing, or false otherwise.