Fix crash when the error signal is raised and glib has G_ENABLE_DEBUG
If glib is built with --enable-debug=yes, it will have G_ENABLE_DEBUG
set internally

This flag changes the g_marshal_value_peek_object() macro to use
g_value_get_object(), which performs sanity checks, instead of doing a
direct offset access.

The definition of our error signal was wrong, using a marshaller with
a GObject in the first parameter but setting the type of the GValue to
G_TYPE_ERROR. Since we have no marshaller for G_TYPE_ERROR, and that all
of those are functionally equivalent (and in fact use the exact same
code in non-debug builds), I went with POINTER for both sides.

The actual crash happened because of a g_return_val_if_fail() in the
sanity checks which made it return NULL after throwing a warning like
this:

    g_value_get_object: assertion 'G_VALUE_HOLDS_OBJECT (value)' failed

This was reported by a gentoo user who had the debug use flag. Thanks!