Fix login hang/timeout when the last page of contacts is empty
The contact list is fetched in pages of 500, but non-friends need to
be filtered out, so you could end up in a situation where the last
page of results is all non-friends. The "contacts" signal was emitted
for each page of results, and has a flag that says if it's complete,
which lets login continue.

So when the last page of the contact list is empty (all non-friends)
the contacts signal isn't emitted, and login gets stuck forever.

This commit changes it so that the signal is emitted with an empty
'users' GSList when it's complete (and when it's an initial fetch,
not when when parsing deltas, which doesn't use the 'users' GSList)

Thanks to dcxk for finding this and helping debug it.