Fix "Failed to read fixed header" with TLS 1.3
Ported from a march 2018 commit in purple-facebook, which is when NSS
updated to TLS 1.3, and purple often uses NSS. Now it's happening with
gnutls 3.6, and bitlbee often uses gnutls.

The code is slightly different because bitlbee has ssl_pending().
Unfortunately, I wasn't able to test this, because even with that
version it works for me. Who knows. Maybe this doesn't fix anything at
all.