Fix taNewMessage bug (#200)
The thrift message on the front of this is not a thrift string, but a
thrift structure. Until recently it was always an empty structure and so
the code was getting away with parsing it as the wrong type.

A few days ago facebook started returning a string in the structure
which broke the parser because it was incorrectly treating the first
byte as a length.